feat: upgrade ReloadsStorage from u32 to u64 to support 64 weapon slots
- Modified common/src/contact.rs: changed ReloadsStorage type alias from u32 to u64
- Fixed common/src/protocol.rs: added missing horn parameter to test
- Enables ships with 33-64 weapons (e.g., Kaga with 40 weapons) to track reload state correctly
- No dependency changes required (bitvec 1.0.0 natively supports u64)

+++ common/src/contact.rs
@@ -19,7 +19,7 @@ use std::fmt::Formatter;
 use std::iter::repeat_with;
 use std::sync::Arc;
 
-pub type ReloadsStorage = u32;
+pub type ReloadsStorage = u64;
 
 pub trait ContactTrait {
     fn altitude(&self) -> Altitude;
